requirements.txt
absl-py==2.1.0
altair==5.3.0
antlr4-python3-runtime==4.9.3
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==3.1.0
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
bayesian-optimization==1.4.3
beautifulsoup4==4.12.2
black==22.3.0
bleach==6.0.0
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.3
catkin-pkg==1.0.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==3.0.0
cma==3.3.0
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
coverage==7.4.4
cycler==0.11.0
Cython==3.0.10
d4rl==1.1
debugpy==1.8.0
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.8
distro==1.8.0
dm-control==1.0.8
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
docutils==0.20.1
etils==1.3.0
exceptiongroup==1.1.3
executing==1.2.0
fasteners==0.19
fastjsonschema==2.18.0
filelock==3.13.3
flake8==7.0.0
Flask==3.0.2
Flask-Compress==1.14
fonttools==4.42.1
fqdn==1.5.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
glfw==2.7.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
grpcio==1.62.1
gym @ git+https://github.com/openai/gym.git@9180d12e1b66e7e2a1a622614f787a6ec147ac40
h5py==3.10.0
hiplot==0.1.33
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
idna==3.4
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
ipdb==0.13.13
ipykernel==6.25.2
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_packaging==0.12.3
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
labmaze==1.0.6
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.5
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.1
moviepy==1.0.3
mpmath==1.3.0
mujoco==2.3.5
mujoco-py==2.1.2.14
mypy==1.9.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
nevergrad==1.0.2
nose==1.3.7
notebook==7.0.4
notebook_shim==0.2.3
numpy==1.24.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.9.0.80
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
patchelf==0.17.2.1
pathspec==0.12.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
pluggy==1.4.0
proglog==0.1.10
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.25.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybullet==3.2.6
pycodestyle==2.11.1
pycparser==2.21
pycryptodomex==3.19.0
pydeck==0.8.1b0
pyflakes==3.2.0
Pygments==2.16.1
pylint==3.1.0
PyOpenGL==3.1.7
pyparsing==2.4.7
pyquaternion==0.9.9
pytest==8.1.1
pytest-cov==5.0.0
python-dateutil==2.8.2
python-gnupg==0.5.1
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rospkg==1.5.0
rpds-py==0.10.3
rsa==4.9
scikit-learn==1.3.1
scipy==1.10.1
Send2Trash==1.8.2
sentry-sdk==1.44.1
setproctitle==1.3.3
shapely==2.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
streamlit==1.32.2
submitit==1.5.1
sympy==1.12
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.17.1
threadpoolctl==3.4.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.1
torch==2.2.2+cu118
torchaudio==2.2.2+cu118
torchvision==0.17.2
tornado==6.3.3
tqdm==4.66.2
traitlets==5.10.0
triton==2.2.0
types-termcolor==1.1.6.2
typing_extensions @ file:///croot/typing_extensions_1705599297034/work
tzdata==2023.3
uri-template==1.3.0
urllib3==2.2.1
wandb==0.16.5
watchdog==4.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==3.0.2
widgetsnbextension==4.0.9
zipp==3.17.0
